pyproject.toml
[tool.poetry]
name = "akiwa"
version = "0.1.0"
description = ""
authors = ["WorldObservationLog <wolc@duck.com>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
graia-broadcast = "^0.23.4"
graia-scheduler = "^0.3.1"
bilibili-api-python = "^16.1.1"
motor = "^3.3.2"
pydantic = "^2.5.3"
websockets = "^10.4"
creart = "^0.3.0"
creart-graia = "^0.1.6"
tomli = "^2.0.1"
graia-saya = "^0.0.19"
beanie = "^1.17.0"
tinydb = "^4.7.1"
jieba = "^0.42.1"
jinja2 = "^3.1.2"
pandas = "^1.5.3"
tenacity = "^8.2.2"
pyppeteer = "^1.0.2"
quart = "^0.19.4"
quart-schema = "^0.18.0"
seaborn = "^0.13.1"
wordcloud = "^1.9.3"



[tool.poetry.group.dev.dependencies]
debugpy = "^1.6.5"
bunnet = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
